diagnosis-api/requirements.txt
psycopg2-binary
jsonschema
Flask
requests
